transcribe_website/transcriber_backend/pyproject.toml
[tool.poetry]
name = "transcriber-backend"
version = "0.1.0"
description = ""
authors = ["BuRny <gamingburny@gmail.com>"]
packages = [
    { include = "src/*.py" },
]

[tool.poetry.dependencies]
python = ">=3.8, <3.12"
# Remote SQL DB
supabase = "^1.0.0"
postgrest-py = "^0.10.2"
# Argparse with dataclass
simple-parsing = "^0.1"
pydantic = "^1.10.5"
toml = "^0.10.2"
faster-whisper = "^0.6.0"
loguru = "^0.7.0"
tqdm = "^4.65.0"
arrow = "^1.2.3"

[tool.poetry.group.dev.dependencies]
# Test library
hypothesis = "^6.8.5"
pytest = "^7.0.0"
pytest-asyncio = "^0.15.1"
# Autoformat
yapf = "^0.32.0"
# Linter
ruff = "^0.0.247"
# Type checker
pyre-check = "^0.9.18"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
based_on_style = "pep8"
column_limit = 120
split_arguments_when_comma_terminated = true
dedent_closing_brackets = true
allow_split_before_dict_value = false

[tool.ruff]
target-version = 'py38'
line-length = 120
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
select = [
    "C4", # flake8-comprehensions
    "E",  # Error
    "F",  # pyflakes
    "BLE",  # flake8-blind-except
    "I",  # isort
    "ISC", # flake8-implicit-str-concat
    "N",  # pep8-naming
    "PGH",  # pygrep-hooks
    "PTH", # flake8-use-pathlib
    "SIM",  # flake8-simplify
    "W",  # Warning
    "Q",  # flake8-quotes
    "YTT", # flake8-2020
    "UP",  # pyupgrade
#    "A",  # flake8-builtins
]

[tool.ruff.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
# Remove once support for py3.8 and 3.9 is dropped
keep-runtime-typing = true

[tool.ruff.pep8-naming]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
classmethod-decorators = ["pydantic.validator", "classmethod"]
